use double in GetLineIntersection

+++ src/ODUtils.cpp
@@ -255,11 +255,11 @@ double getLMT( double ut, double lon )
 // intersect the intersection point may be stored in the floats i_x and i_y.
 bool GetLineIntersection(double p0_x, double p0_y, double p1_x, double p1_y, double p2_x, double p2_y, double p3_x, double p3_y, double* i_x, double* i_y)
 {
-    float s1_x, s1_y, s2_x, s2_y;
+    double s1_x, s1_y, s2_x, s2_y;
     s1_x = p1_x - p0_x;     s1_y = p1_y - p0_y;
     s2_x = p3_x - p2_x;     s2_y = p3_y - p2_y;
     
-    float s, t;
+    double s, t;
     s = (-s1_y * (p0_x - p2_x) + s1_x * (p0_y - p2_y)) / (-s2_x * s1_y + s1_x * s2_y);
     t = ( s2_x * (p0_y - p2_y) - s2_y * (p0_x - p2_x)) / (-s2_x * s1_y + s1_x * s2_y);
